Feat note proportional postion

+++ db/migrate/20200513193359_create_task_lists.rb
@@ -2,8 +2,8 @@ class CreateTaskLists < ActiveRecord::Migration[6.0]
   def change
     create_table :task_lists do |t|
       t.string :title, null: false, default: 'Task list title'
-      t.integer :positionX, null: false, default: 0
-      t.integer :positionY, null: false, default: 0
+      t.float :positionX, precision: 53, null: false, default: 0
+      t.float :positionY, precision:53, null: false, default: 0
       t.integer :width, null: false, default: 0
       t.integer :height, null: false, default: 0
       t.boolean :minimized?, null: false, default: false
